modified apps.py, so overdue borrowings check is now scheduled after migrations and runs once a day
